LayoutTests/fast/scrolling/mac/smooth-scroll-fixed-element-expected.txt

Can't scroll from here after 1 second

PASS window.pageYOffset is 200
PASS successfullyParsed is true

TEST COMPLETE
